Add user defined character classes.
The trickiest part here was getting the union in the Instruction object to work
properly when one of the union members is a unique_ptr. This required carefully
writing constructors/destructions/copy/assignment operations to handle the
unique_ptr based on the type of the instruction.

Ran tests under Valgrind and verified there are no leaks or memory issues.